+++ apps/overstock/components/new-cart/actions.ts
@@ -1,14 +1,25 @@
-'use server'
+"use server";
 
-import { cookies } from "next/headers"
+import { reshapeCart, shopifyFetch } from "lib/shopify";
+import { getCartQuery } from "lib/shopify/queries/cart";
+import type { ShopifyCartOperation } from "lib/shopify/types";
+import { cookies } from "next/headers";
 
 export async function getCart() {
-    const value = cookies().get('cart')?.value
-    console.log(value)
+  const value = cookies().get("cart")?.value;
 
-    if (value) {
-        // TODO: Parse
+  if (value) {
+    const cart = await shopifyFetch<ShopifyCartOperation>({
+      query: getCartQuery,
+      variables: { cartId: `gid://shopify/Cart/${value}` },
+    });
+
+    if (!cart.body.data.cart) {
+      return undefined;
     }
 
-    return undefined
+    return reshapeCart(cart.body.data.cart);
+  }
+
+  return undefined;
 }
